% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_breaks.R
\name{log_breaks}
\alias{log_breaks}
\title{Log Breaks}
\usage{
log_breaks(maj, radix = 10)
}
\arguments{
\item{maj}{Boolean}

\item{radix}{A number, The radix of the log}
}
\description{
This function applies log breaks. It was copied from here:
https://stackoverflow.com/questions/30179442/plotting-minor-breaks-on-a-log-scale-with-ggplot
}
\examples{
 log_breaks(TRUE, 10)
}
