% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_x_log_eng.R
\name{scale_x_log_eng}
\alias{scale_x_log_eng}
\title{X-Log-Axis with Log Breaks}
\usage{
scale_x_log_eng(..., radix = 10)
}
\description{
Applies beautiful log breaks. It was copied from here:
https://stackoverflow.com/questions/30179442/plotting-minor-breaks-on-a-log-scale-with-ggplot
}
\examples{
scale_x_log_eng(radix = 2)
}
