% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tor_theme.R
\name{tor_theme}
\alias{tor_theme}
\title{Torben Theme}
\usage{
tor_theme()
}
\description{
Simply loads Torbens favorite Theme. Simply load like tor_theme
}
\examples{
ggplot(mtcars, aes(mpg, cyl)) +
geom_point() +
tor_theme()
}
